requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
google-generativeai==0.7.2
httpx==0.27.0
pytest==8.2.0